# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Registro Covid API",
        "description": "API do projeto Registro Covid",
        "contact": {
            "email": "email@test.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7000/api/",
            "description": "Registro Covid API"
        }
    ],
    "paths": {
        "/api/corticosteroides": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista corticosteroides",
                "description": "Retorna todos corticosteroides cadastrados no sistema",
                "operationId": "getCorticosteroides",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "descricao": "Prednisona > 40 mg/dia"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/doencas": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista doenças",
                "description": "Retorna todas doenças cadastradas no sistema",
                "operationId": "getDoencas",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "tipo_doenca": 1,
                                            "descricao": "Doença arterial coronariana"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/drogas": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista drogas",
                "description": "Retorna todas drogas cadastrados no sistema",
                "operationId": "getDrogas",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "descricao": "Maconha"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Cadastra drogas",
                "description": "Cadastra droga no sistema",
                "operationId": "storeDrogas",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descricao": {
                                        "description": "Cadastra droga no sistema",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "descricao": "Maconha"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/estados": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista estados",
                "description": "Retorna todos estados cadastrados no sistema",
                "operationId": "getEstados",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "nome": "Acre",
                                            "unidade_federativa": "AC"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/instituicoes": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista instituições",
                "description": "Retorna todas instituições cadastradas no sistema",
                "operationId": "getInstituicoes",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "nome": "Hospital Leonardo da Vinci"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/municipios": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista de municípios",
                "description": "Retorna todos municípios cadastrados no sistema",
                "operationId": "getMunicipios",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1347,
                                            "estado_id": 6,
                                            "nome": "Fortaleza"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/orgaos": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista orgãos",
                "description": "Retorna todos orgãos cadastrados no sistema",
                "operationId": "getOrgaos",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "descricao": "Coração"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/api/situacao-uso-drogas": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista situação uso drogas",
                "description": "Retorna todos as situações de uso de drogas cadastrado no sistema",
                "operationId": "getSituacaoUsoDrogas",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "descricao": "É usuário"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}